examples/requirements.txt
# Additional requirements for examples
openai>=1.0.0
python-dotenv>=1.0.0
aiohttp>=3.8.0